# backend/tests/expected/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AnalyticsOverview": {
        "additionalProperties": true,
        "properties": {
          "lossesBreakdown": {
            "$ref": "#/components/schemas/LossesBreakdown"
          },
          "monthlyBreakdown": {
            "items": {
              "$ref": "#/components/schemas/MonthlyPoint"
            },
            "title": "Monthlybreakdown",
            "type": "array"
          },
          "returnsBreakdown": {
            "$ref": "#/components/schemas/ReturnsBreakdown"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelinePoint"
            },
            "title": "Timeline",
            "type": "array"
          },
          "totalGains": {
            "title": "Totalgains",
            "type": "number"
          },
          "totalLosses": {
            "title": "Totallosses",
            "type": "number"
          },
          "totalReturnRate": {
            "title": "Totalreturnrate",
            "type": "number"
          },
          "trend": {
            "title": "Trend",
            "type": "string"
          }
        },
        "required": [
          "totalReturnRate",
          "trend",
          "returnsBreakdown",
          "lossesBreakdown",
          "totalGains",
          "totalLosses",
          "timeline",
          "monthlyBreakdown"
        ],
        "title": "AnalyticsOverview",
        "type": "object"
      },
...
      "DeletedStatus": {
        "additionalProperties": true,
        "description": "A ``{\"status\": \"deleted\"}`` soft-delete acknowledgement.",
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "DeletedStatus",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InventoryItemCreate": {
        "properties": {
          "acquired_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquired At"
          },
          "markup_paid": {
            "title": "Markup Paid",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tt_value": {
            "title": "Tt Value",
            "type": "number"
          }
        },
        "required": [
          "name",
          "tt_value",
          "markup_paid"
        ],
        "title": "InventoryItemCreate",
        "type": "object"
      },
      "InventoryItemModel": {
        "additionalProperties": true,
        "description": "A persistent-inventory ledger item.",
        "properties": {
          "acquiredAt": {
            "title": "Acquiredat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "markupPaid": {
            "title": "Markuppaid",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "ttValue": {
            "title": "Ttvalue",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "ttValue",
          "markupPaid",
          "acquiredAt"
        ],
        "title": "InventoryItemModel",
        "type": "object"
      },
      "InventoryItemPatch": {
        "properties": {
          "markup_paid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markup Paid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tt_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tt Value"
          }
        },
        "title": "InventoryItemPatch",
        "type": "object"
      },
...
      "LedgerEntryCreate": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "tag": {
            "title": "Tag",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "date",
          "type",
          "description",
          "amount",
          "tag"
        ],
        "title": "LedgerEntryCreate",
        "type": "object"
      },
      "LedgerItem": {
        "additionalProperties": true,
        "description": "A liquid-economy ledger entry (expense or markup).",
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "tag": {
            "title": "Tag",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "date",
          "type",
          "description",
          "amount",
          "tag"
        ],
        "title": "LedgerItem",
        "type": "object"
      },
      "LedgerPresetCreate": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tag": {
            "title": "Tag",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "description",
          "amount",
          "tag"
        ],
        "title": "LedgerPresetCreate",
        "type": "object"
      },
      "LedgerPresetItem": {
        "additionalProperties": true,
        "description": "A ledger quick-entry preset (distinct from a ledger entry: name, not date).",
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tag": {
            "title": "Tag",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "description",
          "amount",
          "tag"
        ],
        "title": "LedgerPresetItem",
        "type": "object"
      },
...
      "SessionDeletedResult": {
        "additionalProperties": true,
        "description": "Acknowledgement that a session was deleted (2-key).",
        "properties": {
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "sessionId"
        ],
        "title": "SessionDeletedResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "EntropiaOrme API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analytics/inventory": {
      "get": {
        "description": "List all inventory items, newest first.",
        "operationId": "list_inventory_items_api_analytics_inventory_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InventoryItemModel"
                  },
                  "title": "Response List Inventory Items Api Analytics Inventory Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Inventory Items",
        "tags": [
          "analytics"
        ]
      },
      "post": {
        "description": "Create a new inventory item.",
        "operationId": "create_inventory_item_api_analytics_inventory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Inventory Item",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/analytics/inventory/{item_id}": {
      "delete": {
        "description": "Hard delete an inventory item (correction path, no ledger entry emitted).",
        "operationId": "delete_inventory_item_api_analytics_inventory__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Inventory Item",
        "tags": [
          "analytics"
        ]
      },
      "patch": {
        "description": "Edit fields on an inventory item. Bumps updated_at.",
        "operationId": "update_inventory_item_api_analytics_inventory__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryItemPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Inventory Item",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/analytics/ledger": {
      "get": {
        "description": "List all ledger entries.\n\nThe activity ledger is liquid economy only. Codex rewards live in\n`codex_claims` and skill-typed quest rewards live in `quest_claims`.\nNew writes never produce PES-tagged ledger rows, so no read-side\nfilter is needed.",
        "operationId": "list_ledger_api_analytics_ledger_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LedgerItem"
                  },
                  "title": "Response List Ledger Api Analytics Ledger Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Ledger",
        "tags": [
          "analytics"
        ]
      },
      "post": {
        "description": "Create a new ledger entry.",
        "operationId": "create_ledger_entry_api_analytics_ledger_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LedgerEntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ledger Entry",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/analytics/ledger/presets": {
      "get": {
        "description": "List all ledger quick-entry presets.",
        "operationId": "list_ledger_presets_api_analytics_ledger_presets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LedgerPresetItem"
                  },
                  "title": "Response List Ledger Presets Api Analytics Ledger Presets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Ledger Presets",
        "tags": [
          "analytics"
        ]
      },
      "post": {
        "description": "Create a new ledger quick-entry preset.",
        "operationId": "create_ledger_preset_api_analytics_ledger_presets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LedgerPresetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerPresetItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ledger Preset",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/analytics/ledger/presets/{preset_id}": {
      "delete": {
        "description": "Delete a ledger quick-entry preset.",
        "operationId": "delete_ledger_preset_api_analytics_ledger_presets__preset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "title": "Preset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Ledger Preset",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/analytics/ledger/{entry_id}": {
      "delete": {
        "description": "Delete a ledger entry.",
        "operationId": "delete_ledger_entry_api_analytics_ledger__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Ledger Entry",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/analytics/overview": {
      "get": {
        "description": "Cross-session sustainability metrics.\n\nTotal Return = (Loot TT + Skill TT + Ledger markup) / (Tracking cost + Ledger expenses).\n?period= all | 30d | 90d | 1y",
        "operationId": "analytics_overview_api_analytics_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analytics Overview",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/demo/analytics/overview": {
      "get": {
        "operationId": "demo_analytics_overview_api_demo_analytics_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Analytics Overview",
        "tags": [
          "demo"
        ]
      }
    },
    "/api/demo/tracking/session/{session_id}": {
      "get": {
        "operationId": "demo_get_session_api_demo_tracking_session__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Get Session",
        "tags": [
          "demo"
        ]
      }
    },
...
    "/api/tracking/session/{session_id}": {
      "delete": {
        "description": "Delete a tracking session and all associated data.",
        "operationId": "delete_session_api_tracking_session__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDeletedResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Session",
        "tags": [
          "tracking"
        ]
      },
      "get": {
        "description": "Get full session detail with aggregated summary.\n\nReturns shape matching the frontend SessionDetail type.",
        "operationId": "get_session_api_tracking_session__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "tracking"
        ]
      }
    },
...
  }
}